refactor: 로또번호들 랜덤 생성하는 기능을 로또에서 로또 상위로 이동
