+++ src/main/java/lotto/domain/Lotto.java
@@ -5,57 +5,32 @@ import java.util.*;
 import static lotto.domain.LottoGameOptions.*;
 
 public class Lotto {
-    private LottoNumbers numbers;
+    private LottoNumbers lottoNumbers;
     private LottoResult result;
 
-    public Lotto() {
-        this(true);
-    }
+    public Lotto(LottoNumbers lottoNumbers) {
+        this.lottoNumbers = lottoNumbers;
 
-    public Lotto(boolean auto) {
-        if (auto) {
-            this.numbers = new LottoNumbers();
-            generateNumbers();
-        }
-    }
-
-    public Lotto(List<Integer> numbers) {
-        this.numbers = new LottoNumbers(numbers);
-
-        if (this.numbers.sizeIsNotSameWith(LOTTO_NUMBER_COUNT)) {
+        if (this.lottoNumbers.sizeIsNotSameWith(LOTTO_NUMBER_COUNT)) {
             throw new IllegalArgumentException();
         }
     }
 
-    private void generateNumbers() {
-        Stack<Integer> wholeLottoNumbers = makeWholeLottoNumbers(LOTTO_NUMBERS_FROM, LOTTO_NUMBERS_TO);
-
-        Collections.shuffle(wholeLottoNumbers);
-
-        while (numbers.sizeIsNotSameWith(LOTTO_NUMBER_COUNT)) {
-            numbers.add(wholeLottoNumbers.pop());
-        }
-    }
-
-    private Stack<Integer> makeWholeLottoNumbers(int start, int end) {
-        Stack<Integer> numbers = new Stack<>();
-        for (int number = start; number <= end; ++number) {
-            numbers.push(number);
-        }
-        return numbers;
+    public Lotto(List<Integer> numbers) {
+        this(new LottoNumbers(numbers));
     }
 
     public String lottoNumberString() {
-        return numbers.formattedString();
+        return lottoNumbers.formattedString();
     }
 
     @Override
     public String toString() {
-        return numbers.toString();
+        return lottoNumbers.toString();
     }
 
     public int numbers(int index) {
-        return numbers.get(index);
+        return lottoNumbers.get(index);
     }
 
 }
