+++ src/test/java/lotto/domain/LottoTest.java
@@ -24,7 +24,7 @@ public class LottoTest {
     @DisplayName("랜덤으로 1~ 45 사이의 중복없는 번호들 생성")
     @Test
     void 구매_로또_번호생성_Test() {
-        Lotto lotto = new Lotto();
+        Lotto lotto = new Lotto(Arrays.asList(1,2,3,4,5,6));
         String[] nums = lotto.lottoNumberString().substring(1, lotto.toString().length()-2).split(", |,");
         for (int i = 0; i < nums.length; ++i) {
             assertThat(Integer.parseInt(nums[i])).isGreaterThanOrEqualTo(1);
